+++ semcore/base-trigger/src/FilterTrigger.jsx
@@ -34,14 +34,10 @@ class RootFilterTrigger extends Component {
             active={active}
             disabled={disabled}
           >
-            {empty ? (
-              <FilterTrigger.Text />
-            ) : (
-              addonTextChildren(Children, FilterTrigger.Text, [
-                FilterTrigger.Addon,
-                FilterTrigger.Counter,
-              ])
-            )}
+            {addonTextChildren(Children, FilterTrigger.Text, [
+              FilterTrigger.Addon,
+              FilterTrigger.Counter,
+            ])}
             {empty && <FilterTrigger.Addon tag={ChevronDown} />}
           </SFilterTrigger>
           {!empty && (