+++ semcore/date-picker/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.3.14] - 2022-10-04
+
+### Changed
+
+- Reverting changes from version `3.3.12` as these changes are implemented in the `ButtonTrigger`.
+
 ## [3.3.13] - 2022-10-06
 
 ### Changed